requirements.txt
azure-ai-textanalytics==5.2.0
faker==37.3.0
openai==1.84.0